cdk-python-tasks/requirements.txt
aws_cdk.aws_ecs_patterns
aws_cdk.aws_ec2
aws_cdk.aws_ecs
aws_cdk.aws_ecr
aws_cdk.aws_ecs_patterns
aws_cdk.aws_servicediscovery
aws_cdk.core